Move truncated projected cost table into independent component
